tidy code
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@57771 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/main/java/org/geogebra/common/euclidian/DrawAxis.java
@@ -889,7 +889,7 @@ public class DrawAxis {
 		g2.drawStringStroke(text, x, y);
 
 		g2.setColor(view.axesColor);
-		g2.drawString(text, (int) (x), (int) y);
+		g2.drawString(text, x, y);
 
 	}
